% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper_roxygen.R
\name{Gabby Freshly}
\title{Boloria dataframe- Gabby Freshly}
\description{
This is the dataframe boloria which records the climate conditions that boloria
chariclea butterfly is living in and the wing length.
