% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{df_apply function}
\usage{
df_apply(.data, .fun, .filter, ...)
}
\description{
when applied to a dataframe, df_apply will a function and a filter.
}
\details{
example: round takes the numeric numbers in a column and rounds them up or down
is.numeric determines if the data in the column is numeric or not
}
