% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\docType{data}
\name{iris_rounded}
\alias{iris_rounded}
\title{try_iris <- df_apply(iris, round, is.numeric)try_iriswhere does it go?}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\usage{
iris_rounded
}
\description{
try_iris <- df_apply(iris, round, is.numeric)try_iriswhere does it go?
}
\keyword{datasets}
